semi/experiment/requirements.txt
tqdm==4.11.2
terminaltables==3.1.0
py-term==0.6